5_RestBrokerData/requirements.txt
asgiref==3.3.4
certifi==2020.12.5
chardet==4.0.0
coverage==5.5
Django==3.2
djangorestframework==3.12.4
gunicorn==20.1.0
idna==2.10
protobuf==3.15.8
psycopg2-binary
pytz==2021.1
requests==2.25.1
six==1.15.0
sqlparse==0.4.1
typing-extensions==3.7.4.3
urllib3==1.26.4
whitenoise==5.2.0
